requirements.txt
awsebcli==3.7.5
blessed==1.9.5
botocore==1.4.10
cement==2.8.2
colorama==0.3.7
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.1
django-appconf==0.6
django-compressor==1.4
djangorestframework==3.0.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
drf-nested-routers==0.9.0
gunicorn==19.1.1
jmespath==0.9.0
pathspec==0.3.4
psycopg2==2.6.1
python-dateutil==2.5.2
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.8.0
static3==0.5.1
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.35.0
